+++ src/core/com/wsscode/pathom/viz/helpers.cljs
@@ -113,7 +113,6 @@
                   :default identity}
       (str/replace s "TaggedValue:" "TaggedValue"))
     (catch :default e
-      (js/console.warn "Safe read failed to read." s e)
       nil)))
 
 (defn safe-read-string
@@ -139,7 +138,7 @@
 
 (defn env-parser-response [env]
   (-> env :result :body
-      (get 'com.wsscode.pathom.viz.client-parser/client-parser-mutation)
+      (get 'com.wsscode.pathom.viz.client-parser/client-parser-request)
       :com.wsscode.pathom.viz.client-parser/client-parser-response))
 
 (defn transform->css [{:keys [x y k] :as t}]
@@ -367,17 +366,11 @@
   IDeref
   (-deref [_o] (-> comp fc/props (get prop)))
 
-  IReset
-  (-reset! [_o new-value] (fm/set-value! comp prop new-value) new-value)
-
-  ISwap
-  (-swap! [a f] (fm/set-value! comp prop (f @a)))
-  (-swap! [a f x] (fm/set-value! comp prop (f @a x)))
-  (-swap! [a f x y] (fm/set-value! comp prop (f @a x y)))
-  (-swap! [a f x y more] (fm/set-value! comp prop (apply f @a x y more))))
+  IFn
+  (-invoke [_o x] (fm/set-value! comp prop x)))
 
 (defn use-component-prop [comp prop]
-  (use-memo (->FulcroComponentProp comp prop) [(-> comp fc/props (get prop))]))
+  (->FulcroComponentProp comp prop))
 
 ;; hooks