+++ tailwind.config.js
@@ -14,6 +14,7 @@ const DEV_PATHS = [
 
 module.exports = {
   mode: 'jit',
+  important: true,
   purge: {
     // in prod look at shadow-cljs output file in dev look at runtime, which will change files that are actually compiled; postcss watch should be a whole lot faster
     content: process.env.NODE_ENV == 'production' ? PROD_PATHS : DEV_PATHS